python-job/requirements.txt
requests
psycopg2-binary==2.8.3
pgcopy
pandas
